feat(mobile): APK Android com MLS nativo (UniFFI+JNA), build validado — autolinking do módulo, cargo-ndk/JDK17/NDK no guia

+++ apps/mobile/modules/janjacord-crypto/android/src/main/java/br/janjacord/crypto/JanjacordCryptoModule.kt
@@ -36,7 +36,7 @@ class JanjacordCryptoModule(reactContext: ReactApplicationContext) :
         safe(promise) { addMember(seedHex, identityId, groupIdHex, keyPackageB64) }
 
     @ReactMethod fun removeMember(seedHex: String, identityId: String, groupIdHex: String, leafIndex: Int, promise: Promise) =
-        safe(promise) { removeMember(seedHex, identityId, groupIdHex, leafIndex) }
+        safe(promise) { removeMember(seedHex, identityId, groupIdHex, leafIndex.toUInt()) }
 
     @ReactMethod fun processCommit(seedHex: String, identityId: String, groupIdHex: String, commitB64: String, promise: Promise) =
         safe(promise) { processCommit(seedHex, identityId, groupIdHex, commitB64); null }
